grpa_client/client-requirement.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.2
chardet==4.0.0
et-xmlfile==1.1.0
idna==3.1
multidict==5.1.0
openpyxl==3.0.7
python-engineio==4.1.0
python-socketio==5.2.1
robotframework==4.0.1
robotframework-excellib==2.0.1
robotframework-pythonlibcore==2.2.1
robotframework-seleniumlibrary==5.1.3
selenium==3.141.0
typing-extensions==3.10.0.0
urllib3==1.26.4
websockets==9.0.1
yarl==1.6.3
